cloud_function/materialize-llm/requirements.txt
google-cloud-storage>=2.10.0
functions-framework>=3.5.0
google-cloud-aiplatform>=1.34.0